refactor: :recycle: Extracting FormsController methods into a concern
